+++ src/fields/Tags.php
@@ -44,7 +44,7 @@ class Tags extends BaseRelationField
     /**
      * @inheritdoc
      */
-    protected static function elementType(): string
+    public static function elementType(): string
     {
         return Tag::class;
     }